https://online.sbis.ru/opendoc.html?guid=6eb01729-d006-4c90-8c60-36306fa3ddad Добавил дополнительную демку для проверки работы tagStyle через параметры шаблона
