+++ Controls-demo/grid/TagStyle/TagStyle.wml
@@ -1,10 +1,21 @@
-<div class="controlsDemo__wrapper controlsDemo__grid-tagStyle">
-   <Controls.grid:View
-      keyProperty="id"
-      source="{{_viewSource}}"
-      columns="{{_columns}}">
-      <ws:itemTemplate>
-         <ws:partial template="Controls/grid:ItemTemplate" highlightOnHover="{{true}}"/>
-      </ws:itemTemplate>
-   </Controls.grid:View>
+<div class="controlsDemo__wrapper controlsDemo__flexRow">
+
+   <div class="controlsDemo__cell">
+      <Router.router:Reference attr:class="controls-text-label_theme-{{_options.theme}}"
+                               state="/Controls-demo/app/:app"
+                               app="{{ 'Controls-demo/grid/TagStyle/TagStyleFromCellData/Index' }}">
+         <a href="{{content.href}}">Из данных в ячейке</a>
+      </Router.router:Reference>
+      <Controls-demo.grid.TagStyle.TagStyleFromCellData.Index/>
+   </div>
+
+   <div class="controlsDemo__cell">
+      <Router.router:Reference attr:class="controls-text-label_theme-{{_options.theme}}"
+                               state="/Controls-demo/app/:app"
+                               app="{{ 'Controls-demo/grid/TagStyle/TagStyleFromTemplateParam/Index' }}">
+         <a href="{{content.href}}">Из параметра шаблона для колонки</a>
+      </Router.router:Reference>
+      <Controls-demo.grid.TagStyle.TagStyleFromTemplateParam.Index/>
+   </div>
+
 </div>
