+++ Controls-demo/grid/TagStyle/TagStyleFromCellData/TagStyleFromCellData.wml
@@ -1,0 +1,7 @@
+<div class="controlsDemo__wrapper controlsDemo__grid-tagStyle">
+   <Controls.grid:View
+      keyProperty="id"
+      source="{{_viewSource}}"
+      columns="{{_columns}}">
+   </Controls.grid:View>
+</div>
